Refactor SeedData to use async/await pattern and improve logging
- Converted EnsureSeedData method to EnsureSeedDataAsync for asynchronous execution.
- Introduced separate methods for seeding Alice, Bob, System Admin, Admin, and Sellers for better organization and readability.
- Added logging improvements to provide more context during the seeding process.
- Used GUIDs for Alice and Bob to ensure consistent user IDs.
- Implemented a new structure for seeding sellers with a loop for better scalability.
- Enhanced error handling and logging for claims and address creation.

+++ libs/HiveSpace.Infrastructure.Messaging.Shared/Events/Products/ProductSkuUpdatedEvent.cs
@@ -7,6 +7,6 @@ public record ProductSkuUpdatedEvent(
     int SkuId,
     string SkuNo,
     int Quantity,
-    decimal Price,
+    long Price,
     string Currency
 ) : IntegrationEvent;
